third_party/WebKit/LayoutTests/fast/backgrounds/background-shorthand-multiple-layers.html

<!DOCTYPE html>
<html>
<style>#test { background: url(dummy://foo), red url(dummy://bar); }</style>
<script src="../../resources/js-test.js"></script>
<div id="test"> </div>
<script>
description("Tests that serialization of background shorthand property with multiple layers is correct.");

var e = document.getElementById('test');
shouldBeEqualToString("getComputedStyle(e).background", 'url("dummy://foo") repeat scroll 0% 0% / auto padding-box border-box, rgb(255, 0, 0) url("dummy://bar") repeat scroll 0% 0% / auto padding-box border-box')

</script>
</body>
</html>
